FragPool works
- oz
  * Config::clear behaviour flipped
  * System::bell()
  * Math::normalRand(), Math::centralRand()
- matrix
  * changed frag damage calculation
  * Struct::id removed
- client
  * FragPool works
  * Context::drawFrag, fragPools added to Context
  * FragPool models loaded/unloaded by Loader
  * Imago ctor has Object* parameter

+++ src/oz/BufferStream.hpp
@@ -15,7 +15,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * Davorin Učakar <davorin.ucakar@gmail.com>
+ * Davorin Učakar
+ * <davorin.ucakar@gmail.com>
  */
 
 /**
@@ -155,7 +156,7 @@ class BufferStream
     OZ_ALWAYS_INLINE
     int length() const
     {
-      return int( size_t( pos - start ) );
+      return int( pos - start );
     }
 
     /**
@@ -164,7 +165,7 @@ class BufferStream
     OZ_ALWAYS_INLINE
     int capacity() const
     {
-      return int( size_t( end - start ) );
+      return int( end - start );
     }
 
     /**
@@ -173,7 +174,7 @@ class BufferStream
     OZ_ALWAYS_INLINE
     int available() const
     {
-      return int( size_t( end - pos ) );
+      return int( end - pos );
     }
 
     /**
